move safe_sqr and _approximate_mode into extmath
